API/requirements.txt
Flask==2.3.3
requests==2.31.0
google-cloud-bigquery==3.8.0
google-auth==2.21.0
pandas
db-dtypes